tests/wpt/tests/css/css-scoping/host-in-host-selector.html

<!DOCTYPE html>
<title>CSS Scoping: :host(:host)</title>
<link rel="help" href="https://drafts.csswg.org/css-scoping/#host-selector">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="host">
  <template shadowrootmode="open">
    <style>
      /* Should not match */
      :host(:host) {
        --x:FAIL;
      }
    </style>
  </template>
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(host).getPropertyValue('--x'), '');
  });
</script>
